+++ abc-bank/Account.cs
@@ -18,7 +18,7 @@ namespace abc_bank
             transactions = new List<Transaction>();
         }
 
-        public void Deposit(double amount) 
+        public void Deposit(decimal amount) 
         {
             if (amount <= 0) {
                 throw new ArgumentException("amount must be greater than zero");
@@ -27,7 +27,7 @@ namespace abc_bank
             }
         }
 
-        public void Withdraw(double amount) 
+        public void Withdraw(decimal amount) 
         {
             if (amount <= 0) {
                 throw new ArgumentException("amount must be greater than zero");
@@ -36,31 +36,31 @@ namespace abc_bank
             }
         }
 
-        public double InterestEarned() 
+        public decimal InterestEarned() 
         {
-            double amount = sumTransactions();
+            decimal amount = sumTransactions();
             switch(_accountType){
                 case AccountType.SAVINGS:
                     if (amount <= 1000)
-                        return amount * 0.001;
+                        return amount * 0.001m;
                     else
-                        return 1 + (amount-1000) * 0.002;
+                        return 1 + (amount-1000) * 0.002m;
     //            case SUPER_SAVINGS:
     //                if (amount <= 4000)
     //                    return 20;
                 case AccountType.MAXI_SAVINGS:
                     if (amount <= 1000)
-                        return amount * 0.02;
+                        return amount * 0.02m;
                     if (amount <= 2000)
-                        return 20 + (amount-1000) * 0.05;
-                    return 70 + (amount-2000) * 0.1;
+                        return 20 + (amount-1000) * 0.05m;
+                    return 70 + (amount-2000) * 0.1m;
                 default:
-                    return amount * 0.001;
+                    return amount * 0.001m;
             }
         }
 
-        public double sumTransactions() {
-            double amount = 0.0;
+        public decimal sumTransactions() {
+            decimal amount = 0.0m;
             foreach (Transaction t in transactions)
                 amount += t.Amount;
             return amount;
